Move expanded_column_config to columns
