+++ corehq/apps/userreports/sql/__init__.py
@@ -1,3 +1,2 @@
 from .adapter import get_indicator_table, IndicatorSqlAdapter, metadata, ErrorRaisingIndicatorSqlAdapter
-from .columns import get_expanded_column_config
 from .util import get_column_name
