+++ corehq/apps/userreports/sql/columns.py
@@ -1,8 +1,5 @@
 import sqlalchemy
 from sqlagg import SumWhen
-from django.utils.translation import ugettext as _
-from corehq.apps.userreports.columns import ColumnConfig, get_distinct_values
-from corehq.apps.userreports.exceptions import ColumnNotFoundError
 from corehq.apps.reports.sqlreport import DatabaseColumn
 from fluff import TYPE_STRING
 from fluff.util import get_column_type
@@ -30,47 +27,7 @@ def column_to_sql(column):
     )
 
 
-def get_expanded_column_config(data_source_configuration, column_config, lang):
-    """
-    Given an ExpandedColumn, return a list of DatabaseColumn objects. Each DatabaseColumn
-    is configured to show the number of occurrences of one of the values present for
-    the ExpandedColumn's field.
-
-    This function also adds warnings to the column_warnings parameter.
-
-    :param data_source_configuration:
-    :param column_config:
-    :param column_warnings:
-    :return:
-    """
-    column_warnings = []
-    try:
-        vals, over_expansion_limit = get_distinct_values(
-            data_source_configuration, column_config, column_config.max_expansion
-        )
-    except ColumnNotFoundError as e:
-        return ColumnConfig([], warnings=[unicode(e)])
-    else:
-        if over_expansion_limit:
-            column_warnings.append(_(
-                u'The "{header}" column had too many values to expand! '
-                u'Expansion limited to {max} distinct values.'
-            ).format(
-                header=column_config.get_header(lang),
-                max=column_config.max_expansion,
-            ))
-        return ColumnConfig(_expand_column(column_config, vals, lang), warnings=column_warnings)
-
-
-def _expand_column(report_column, distinct_values, lang):
-    """
-    Given an ExpandedColumn, return a list of DatabaseColumn objects. Each column
-    is configured to show the number of occurrences of one of the given distinct_values.
-
-    :param report_column:
-    :param distinct_values:
-    :return:
-    """
+def expand_column(report_column, distinct_values, lang):
     columns = []
     for index, val in enumerate(distinct_values):
         alias = u"{}-{}".format(report_column.column_id, index)
